Fix totem of dawn cleanse particles.
Previously, they were spawning somewhere random (and disconnected from the
spot actually being cleansed, which was controlled by the server).  Now we
spawn cleanse particles with a packet sent from the server.

We limit the packets to only send when the totem is "caught up" to prevent
a spam of packets when a totem is loaded.  I also changed the particle
behavior to spawn 50 smoke particles at the height level near the player
instead of 1000 smoke particles across the entire column.  Realistically
this isn't as huge a perf improvement as it sounds like because smoke
particles only spawn within 16 blocks of a player.